dom/html/test/test_option_defaultSelected.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=927796
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 927796</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=927796">Mozilla Bug 927796</a>
<p id="display">
<select id="s1">
  <option selected>one</option>
  <option>two</option>
</select>
<select id="s2" size="5">
  <option selected>one</option>
  <option>two</option>
</select>
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 927796 **/
  var s1 = $("s1");
  s1.options[0].defaultSelected = false;
  is(s1.options[0].selected, true,
     "First option in combobox should still be selected");
  is(s1.options[1].selected, false,
     "Second option in combobox should not be selected");

  var s2 = $("s2");
  s2.options[0].defaultSelected = false;
  is(s2.options[0].selected, false,
     "First option in listbox should not be selected");
  is(s2.options[1].selected, false,
     "Second option in listbox should not be selected");
  </script>
</body>
</html>
